raptor now uses beast version 10.0-RC1 in an attempt to get raptor runner to run
