+++ benchmarking/test/beast_raven_raptor.json
@@ -1,0 +1,68 @@
+[
+    {
+        "name": "Beast ",
+        "iterations": 5,
+        "command": {
+            "path": "bash",
+            "args": [
+                "/opt/bitnami/spark/beast-0.10.0-RC1/bin/beast",
+                "--class",
+                "dk.itu.raven.beast.App",
+                "/home/beast-uber-examples-0.9.4-SNAPSHOT.jar",
+                "-iv",
+                "/home/boundaries.zip",
+                "-ir",
+                "/home/glc2000"
+            ]
+        },
+        "environment_options": {
+            "environment_type": "docker",
+            "docker_file_path": "./runners/beast-runner/",
+            "docker_mount_path": "\\\\wsl.localhost\\Ubuntu\\home\\johan\\mount"
+        }
+    },
+    {
+        "name": "Parallel Raven",
+        "iterations": 25,
+        "command": {
+            "path": "java",
+            "args": [
+                "-jar",
+                "/home/raven-runner-TM.jar",
+                "-iv",
+                "/home/boundaries/ne_10m_admin_0_countries.shp",
+                "-ir",
+                "/home/glc2000",
+                "-t",
+                "parallel",
+                "-ts",
+                "2048"
+            ]
+        },
+        "environment_options": {
+            "environment_type": "docker",
+            "docker_file_path": "./runners/raven-runner/",
+            "docker_mount_path": "\\\\wsl.localhost\\Ubuntu\\home\\johan\\mount"
+        }
+    },
+    {
+        "name": "Raptor",
+        "iterations": 25,
+        "command": {
+            "path": "java",
+            "args": [
+                "-jar",
+                "/home/raptor-runner-1.0.jar",
+                "-iv",
+                "/home/boundaries.zip",
+                "-ir",
+                "/home/glc2000"
+            ]
+        },
+        "environment_options": {
+            "environment_type": "docker",
+            "docker_file_path": "./runners/raptor-runner/",
+            "docker_mount_path": "\\\\wsl.localhost\\Ubuntu\\home\\johan\\mount"
+        }
+    }
+]
