+++ raven/pom.xml
@@ -71,11 +71,6 @@
 			<artifactId>rtree2</artifactId>
 			<version>0.9.3</version>
 		</dependency>
-		<dependency>
-			<groupId>mil.nga</groupId>
-			<artifactId>tiff</artifactId>
-			<version>2.0.5</version>
-		</dependency>
 		<!--Parse
 		commandline arguments-->
 		<dependency>
